refactor: Update import path for useFetchSearchLecture hook
